Add skills UI, pull out realm auth code into common class
